infrastructure/server/proxy-all.html

<!DOCTYPE HTML>
<title>test behavior of PROXY configuration (PAC)</title>
<meta name="proxy" content="all">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    promise_test(async t => {
        let response = null;
        try {
            response = await fetch('http://not-a-real-domain.wpt.test/infrastructure/resources/ok.txt');
        } catch (e) {
            assert_unreached("Response to arbitrary domain should succeed");
        }
        const text = await response.text();
        assert_equals(text, 'OK');
    }, 'test that meta with proxy=all is respected');
</script>
